Enable output logs for qsub, stderr, stdin of the submitted job
In order to get all the log files, some reconstruction was needed, e.g. the
workload generator object is now moved inside the connection object. This
removes dependency inside the run.py and improves the encapsulation of the
connection object.

The output logs are collected by the connection as well and will be printed
to the info logger.

Debug output has been reduced, due to the almost complete implementation.

Change-Id: Idb773018b166f7ad8e08677d2af9c0692257b4a9
Signed-off-by: Uwe Schilling <uwe.schilling@hlrs.de>